+++ cpp/libdictation-client/dictation_client.cpp
@@ -23,15 +23,15 @@ namespace {
     }
 
 
-    std::shared_ptr<grpc::ChannelCredentials> create_channel_credentials(const std::string& ssl_directory) {
-        if (ssl_directory.empty())
+    std::shared_ptr<grpc::ChannelCredentials> create_channel_credentials(const std::string& tls_directory) {
+        if (tls_directory.empty())
         {
             return grpc::InsecureChannelCredentials();
         }
 
-        std::string cert = read_file(ssl_directory + "/client.crt");
-        std::string key = read_file(ssl_directory + "/client.key");
-        std::string root = read_file(ssl_directory + "/ca.crt");
+        std::string cert = read_file(tls_directory + "/client.crt");
+        std::string key = read_file(tls_directory + "/client.key");
+        std::string root = read_file(tls_directory + "/ca.crt");
         grpc::SslCredentialsOptions opts = {root, key, cert};
 
         return grpc::SslCredentials(opts);
@@ -55,7 +55,7 @@ gsapi::RecognizeResponse DictationClient::Recognize(DictationSessionConfig& conf
 
     gsapi::RecognizeResponse response;
 
-    auto stub = gsapi::Speech::NewStub(grpc::CreateChannel(service_address_, create_channel_credentials(ssl_directory_)));
+    auto stub = gsapi::Speech::NewStub(grpc::CreateChannel(service_address_, create_channel_credentials(tls_directory_)));
 
     grpc::ClientContext context;
     prepare_context(context, config);
@@ -71,7 +71,7 @@ gsapi::RecognizeResponse DictationClient::Recognize(DictationSessionConfig& conf
 
 
 std::vector<gsapi::StreamingRecognizeResponse> DictationClient::StreamingRecognize(DictationSessionConfig& config, const WAV_DATA& wav_data) const {
-    auto stub = gsapi::Speech::NewStub(grpc::CreateChannel(service_address_, create_channel_credentials(ssl_directory_)));
+    auto stub = gsapi::Speech::NewStub(grpc::CreateChannel(service_address_, create_channel_credentials(tls_directory_)));
 
     grpc::ClientContext context;
     prepare_context(context, config);